% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlog.R
\name{z}
\alias{z}
\alias{zlog}
\title{Calculate z/zlog values}
\usage{
z(x, limits, probs = c(0.025, 0.975))

zlog(x, limits, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{\code{numeric}, laboratory values}

\item{limits}{\code{numeric}, lower and upper reference limits.}

\item{probs}{\code{numeric}, probabilities of the lower and upper reference limit,
default: \code{c(0.025, 0.975)} (spanning 95 \%).}
}
\value{
\code{numeric}, z or z(log) values.
}
\description{
Calculates the z or z(log) values for laboratory measurement standardisation
as proposed in Hoffmann 2017 et al.
}
\details{
The z values is calculated as follows (assuming that the limits where 0.025
and 0.975 quantiles):
\eqn{z = (x - (limits_1 + limits_2 )/2) * 3.92/(limits_2 - limits_1)}.
}
\author{
Sebastian Gibb
}
